+++ bazelrunner/src/main/java/org/jetbrains/bsp/bazel/bazelrunner/BazelRunnerBuilder.java
@@ -1,5 +1,6 @@
 package org.jetbrains.bsp.bazel.bazelrunner;
 
+import ch.epfl.scala.bsp4j.BuildTargetIdentifier;
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
@@ -55,6 +56,14 @@ public class BazelRunnerBuilder {
 
   public BazelRunnerBuilder withTargets(List<String> bazelTargets) {
     String joinedTargets = BazelArgumentsUtils.getJoinedBazelTargets(bazelTargets);
+    arguments.add(joinedTargets);
+
+    return this;
+  }
+
+  public BazelRunnerBuilder withTargets(
+      List<BuildTargetIdentifier> includedTargets, List<BuildTargetIdentifier> excludedTargets) {
+    var joinedTargets = BazelArgumentsUtils.joinBazelTargets(includedTargets, excludedTargets);
     arguments.add(joinedTargets);
 
     return this;